+++ src/machine/operation/jump_if_false.rs
@@ -13,7 +13,7 @@ impl JumpIfFalse {
 }
 
 impl Operation for JumpIfFalse {
-    fn execute(&self, registry: &mut Vec<i32>) -> OpResult {
+    fn execute(&self, registry: &mut Vec<i64>) -> OpResult {
         let x = registry.get(0);
         let out_param = registry.get(1);
